# docs/swagger.yaml
basePath: /api
definitions:
...
  ds.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/ds.UserDTO'
    type: object
...
  ds.UserDTO:
    properties:
      full_name:
        type: string
      id:
        type: integer
      moderator:
        type: boolean
      username:
        type: string
    type: object
  ds.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  ds.UserRegisterRequest:
    properties:
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - full_name
    - password
    - username
    type: object
  ds.UserUpdateRequest:
    properties:
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления нагрузками и сессиями расчета в системе.
  title: API для системы расчета нагрузок
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Получение JWT токена по логину и паролю для доступа к защищенным
        эндпоинтам.
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/ds.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.LoginResponse'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверные учетные данные
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Аутентификация пользователя (все)
      tags:
      - auth
  /auth/logout:
    post:
      description: Добавляет текущий JWT токен в черный список, делая его недействительным.
        Требует авторизации.
      responses:
        "200":
          description: Сообщение об успехе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход из системы (авторизованный пользователь)
      tags:
      - auth
...
  /users:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе. По умолчанию роль "пользователь",
        не "модератор".
      parameters:
      - description: Данные для регистрации
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/ds.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.UserDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя (все)
      tags:
      - auth
  /users/{id}:
    get:
      description: Возвращает публичные данные пользователя. Требует авторизации.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.UserDTO'
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получение данных пользователя по ID (авторизованный пользователь)
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет имя пользователя или пароль. Требует авторизации.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/ds.UserUpdateRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновление данных пользователя (авторизованный пользователь)
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"